ADD RoutePageGenerator::update() organize by path
Modify RoutePageGenerator in order to organize pages by path: it assumes that if a page path if a prefix for an other page path, then the first page is an ancestor of the second page.

This makes the page tree more comfortable to read for end users and makes me able to manage breadcrumbs and menus using the natural parent-child relationship between pages.